+++ R/ipw_did_panel.R
@@ -91,14 +91,7 @@ ipw_did_panel <-function(y1, y0, D, covariates, i.weights = NULL,
   #-----------------------------------------------------------------------------
   #Pscore estimation (logit) and also its fitted values
   #PS <- suppressWarnings(stats::glm(D ~ -1 + int.cov, family = "binomial", weights = i.weights))
-  PS <- suppressWarnings(fastglm::fastglm(x = int.cov,
-                                          y = D,
-                                          family = stats::binomial(),
-                                          weights = i.weights,
-                                          intercept = FALSE,
-                                          method = 3
-  ))
-  class(PS) <- "glm" #this allow us to use vcov
+  PS <- suppressWarnings(fastglm_fit(int.cov, D, stats::binomial(), i.weights, method = 3))
   if(PS$converged == FALSE){
     warning("Propensity score estimation did not converge.")
   }
@@ -132,15 +125,18 @@ ipw_did_panel <-function(y1, y0, D, covariates, i.weights = NULL,
   # Asymptotic linear representation of logit's beta's
   score.ps <- i.weights * (D - ps.fit) * int.cov
   #Hessian.ps <- stats::vcov(PS) * n
-  Hessian.ps <- chol2inv(chol(t(int.cov) %*% (W * int.cov))) * n
+  XtWX.ps <- base::crossprod(int.cov, W * int.cov)
+  if (base::rcond(XtWX.ps) < .Machine$double.eps) {
+    stop("The propensity score design matrix is singular. Consider removing some covariates.")
+  }
+  Hessian.ps <- chol2inv(chol(XtWX.ps)) * n
   asy.lin.rep.ps <-  score.ps %*% Hessian.ps
   #-----------------------------------------------------------------------------
   # Now, get the influence function of control component
   # Leading term of the influence function: no estimation effect
   att.lin1 <- att.treat - att.cont
   # Derivative matrix (k x 1 vector)
-  mom.logit <- att.cont * int.cov
-  mom.logit <- colMeans(mom.logit)
+  mom.logit <- as.vector(base::crossprod(att.cont, int.cov))/n
   # Now the influence function related to estimation effect of pscores
   att.lin2 <- asy.lin.rep.ps %*% mom.logit
   #get the influence function of the DR estimator (put all pieces together)