+++ R/reg_did_panel.R
@@ -108,12 +108,10 @@ reg_did_panel <-function(y1, y0, D, covariates, i.weights = NULL,
   #                                    subset = D==0,
   #                                    weights = i.weights))
   control_filter <- (D == 0)
-  reg.coeff <- stats::coef(fastglm::fastglm(
-                            x = int.cov[control_filter, , drop = FALSE],
-                            y = deltaY[control_filter],
-                            weights = i.weights[control_filter],
-                            family = gaussian(link = "identity")
-  ))
+  reg.coeff <- fastglm_fit(int.cov[control_filter, , drop = FALSE],
+                           deltaY[control_filter],
+                           gaussian(link = "identity"),
+                           i.weights[control_filter])$coefficients
   if(anyNA(reg.coeff)){
     stop("Outcome regression model coefficients have NA components. \n Multicollinearity (or lack of variation) of covariates is probably the reason for it.")
   }
@@ -157,7 +155,7 @@ reg_did_panel <-function(y1, y0, D, covariates, i.weights = NULL,
   inf.cont.1 <- (reg.att.cont - w.cont * eta.cont)
   # Estimation effect from beta hat (OLS using only controls)
   # Derivative matrix (k x 1 vector)
-  M1 <- base::colMeans(w.cont * int.cov)
+  M1 <- as.vector(base::crossprod(w.cont, int.cov))/n
   # Now get the influence function related to the estimation effect related to beta's
   inf.cont.2 <- asy.lin.rep.ols %*% M1
   # Influence function for the control component