+++ R/wboot_drdid_rc1.R
@@ -10,12 +10,7 @@ wboot_drdid_rc1 <- function(nn, n, y, post, D, int.cov, i.weights,
   b.weights <- as.vector(i.weights * v)
   # Propensity score estimation
   # ps.b <- suppressWarnings(stats::glm(D ~ -1 + int.cov, family = "binomial", weights = b.weights)$fitted.values)
-  ps.b <- suppressWarnings(fastglm::fastglm(x = int.cov,
-                                            y = D,
-                                            family = stats::binomial(),
-                                            weights =  b.weights,
-                                            intercept = FALSE,
-                                            method = 3)$fitted.values)
+  ps.b <- suppressWarnings(fastglm_fit(int.cov, D, stats::binomial(), b.weights, method = 3)$fitted.values)
   ps.b <- as.vector(ps.b)
   ps.b <- pmin(ps.b, 1 - 1e-6)
   trim.ps <- (ps.b < 1.01)
@@ -25,24 +20,20 @@ wboot_drdid_rc1 <- function(nn, n, y, post, D, int.cov, i.weights,
   #                                        subset = ((D==0) & (post==0)),
   #                                        weights = b.weights))
   control_pre <- (D == 0) & (post == 0)
-  reg.coeff.pre.b <- stats::coef(fastglm::fastglm(
-                                  x = int.cov[control_pre, , drop = FALSE],
-                                  y = y[control_pre],
-                                  weights = b.weights[control_pre],
-                                  family = gaussian(link = "identity")
-  ))
+  reg.coeff.pre.b <- fastglm_fit(int.cov[control_pre, , drop = FALSE],
+                                 y[control_pre],
+                                 gaussian(link = "identity"),
+                                 b.weights[control_pre])$coefficients
   out.y.pre.b <-   as.vector(tcrossprod(reg.coeff.pre.b, int.cov))
   #Compute the Outcome regression for the control group at the pre-treatment period, using ols.
   # reg.coeff.post.b <- stats::coef(stats::lm(y ~ -1 + int.cov,
   #                                         subset = ((D==0) & (post==1)),
   #                                         weights = b.weights))
   control_post <- (D == 0) & (post == 1)
-  reg.coeff.post.b <- stats::coef(fastglm::fastglm(
-                                  x = int.cov[control_post, , drop = FALSE],
-                                  y = y[control_post],
-                                  weights = b.weights[control_post],
-                                  family = gaussian(link = "identity")
-  ))
+  reg.coeff.post.b <- fastglm_fit(int.cov[control_post, , drop = FALSE],
+                                  y[control_post],
+                                  gaussian(link = "identity"),
+                                  b.weights[control_post])$coefficients
   out.y.post.b <-   as.vector(tcrossprod(reg.coeff.post.b, int.cov))
   # Combine the ORs
   out.y.b <- post * out.y.post.b + (1 - post) * out.y.pre.b